GBX.NET 2.0.7

+++ Src/GBX.NET/Serialization/GbxWriter.cs
@@ -878,7 +878,7 @@ public sealed partial class GbxWriter : BinaryWriter, IGbxWriter
                 Write(value);
                 break;
             case >= byte.MaxValue:
-                Write((uint)value);
+                Write((ushort)value);
                 break;
             default:
                 Write((byte)value);
